Update DefaultEntities.php
CROSS_SELLING_ACCESSORY and CROSS_SELLING_SIMILAR are set to cross_selling_accessory_item and cross_selling_similar_item and used in CrossSellingConverter.php with `if ($data['type'] === DefaultEntities::CROSS_SELLING_SIMILAR)` 
However $data['type'] is product_cross_selling_accessory or product_cross_selling_similar so this if query never works. 
therefore the naming has to be changed for it to work. 

This is related to https://issues.shopware.com/issues/MIG-274